[mob] Refactor
