+++ mobile/lib/emergency/emergency_service.dart
@@ -9,9 +9,9 @@ import "package:photos/core/configuration.dart";
 import "package:photos/core/network/network.dart";
 import "package:photos/emergency/model.dart";
 import "package:photos/generated/l10n.dart";
+import "package:photos/models/api/user/key_attributes.dart";
+import "package:photos/models/api/user/set_keys_request.dart";
 import "package:photos/models/api/user/srp.dart";
-import "package:photos/models/key_attributes.dart";
-import "package:photos/models/set_keys_request.dart";
 import "package:photos/services/user_service.dart";
 import "package:photos/ui/common/user_dialogs.dart";
 import "package:photos/utils/crypto_util.dart";
