+++ mobile/lib/emergency/recover_others_account.dart
@@ -7,8 +7,8 @@ import 'package:password_strength/password_strength.dart';
 import "package:photos/emergency/emergency_service.dart";
 import "package:photos/emergency/model.dart";
 import "package:photos/generated/l10n.dart";
-import "package:photos/models/key_attributes.dart";
-import "package:photos/models/set_keys_request.dart";
+import "package:photos/models/api/user/key_attributes.dart";
+import "package:photos/models/api/user/set_keys_request.dart";
 import 'package:photos/ui/common/dynamic_fab.dart';
 import "package:photos/utils/crypto_util.dart";
 import 'package:photos/utils/dialog_util.dart';
