+++ mobile/lib/models/api/collection/collection_file_item.dart
