+++ mobile/lib/models/api/collection/trash_item_request.dart
