+++ mobile/lib/models/api/entity/data.dart
@@ -1,6 +1,6 @@
 import 'dart:convert';
 
-import 'package:flutter/material.dart';
+import "package:freezed_annotation/freezed_annotation.dart";
 
 @immutable
 class EntityData {
