+++ mobile/lib/models/user_details.dart
@@ -2,9 +2,9 @@ import 'dart:convert';
 import 'dart:math';
 
 import 'package:collection/collection.dart';
+import "package:photos/models/api/billing/subscription.dart";
 import "package:photos/models/api/storage_bonus/bonus.dart";
 import 'package:photos/models/file/file_type.dart';
-import 'package:photos/models/subscription.dart';
 
 class UserDetails {
   final String email;
