+++ mobile/lib/module/upload/model/upload_url.dart
