+++ mobile/lib/services/billing_service.dart
@@ -8,8 +8,8 @@ import 'package:in_app_purchase/in_app_purchase.dart';
 import 'package:logging/logging.dart';
 import 'package:photos/core/errors.dart';
 import "package:photos/generated/l10n.dart";
-import 'package:photos/models/billing_plan.dart';
-import 'package:photos/models/subscription.dart';
+import 'package:photos/models/api/billing/billing_plan.dart';
+import 'package:photos/models/api/billing/subscription.dart';
 import 'package:photos/models/user_details.dart';
 import 'package:photos/services/user_service.dart';
 import 'package:photos/ui/common/web_page.dart';
