+++ mobile/lib/services/trash_sync_service.dart
@@ -9,10 +9,10 @@ import 'package:photos/events/collection_updated_event.dart';
 import 'package:photos/events/force_reload_trash_page_event.dart';
 import 'package:photos/events/trash_updated_event.dart';
 import 'package:photos/extensions/list.dart';
+import 'package:photos/models/api/collection/trash_item_request.dart';
 import 'package:photos/models/file/file.dart';
 import 'package:photos/models/file/trash_file.dart';
 import 'package:photos/models/ignored_file.dart';
-import 'package:photos/models/trash_item_request.dart';
 import 'package:photos/services/ignored_files_service.dart';
 import 'package:photos/utils/trash_diff_fetcher.dart';
 import 'package:shared_preferences/shared_preferences.dart';
