+++ mobile/lib/ui/account/delete_account_page.dart
@@ -6,7 +6,7 @@ import 'package:flutter/material.dart';
 import "package:logging/logging.dart";
 import 'package:photos/core/configuration.dart';
 import "package:photos/generated/l10n.dart";
-import 'package:photos/models/delete_account.dart';
+import 'package:photos/models/api/user/delete_account.dart';
 import 'package:photos/services/user_service.dart';
 import 'package:photos/theme/ente_theme.dart';
 import 'package:photos/ui/components/buttons/button_widget.dart';
