+++ mobile/lib/ui/payment/payment_web_page.dart
@@ -8,7 +8,7 @@ import 'package:logging/logging.dart';
 import "package:photos/core/constants.dart";
 import 'package:photos/ente_theme_data.dart';
 import "package:photos/generated/l10n.dart";
-import 'package:photos/models/subscription.dart';
+import 'package:photos/models/api/billing/subscription.dart';
 import "package:photos/service_locator.dart";
 import 'package:photos/services/billing_service.dart';
 import 'package:photos/services/user_service.dart';
