+++ mobile/lib/ui/payment/store_subscription_page.dart
@@ -9,8 +9,8 @@ import 'package:photos/core/errors.dart';
 import 'package:photos/core/event_bus.dart';
 import 'package:photos/events/subscription_purchased_event.dart';
 import "package:photos/generated/l10n.dart";
-import 'package:photos/models/billing_plan.dart';
-import 'package:photos/models/subscription.dart';
+import 'package:photos/models/api/billing/billing_plan.dart';
+import 'package:photos/models/api/billing/subscription.dart';
 import 'package:photos/models/user_details.dart';
 import "package:photos/service_locator.dart";
 import 'package:photos/services/user_service.dart';
