+++ mobile/lib/ui/payment/subscription_common_widgets.dart
@@ -2,8 +2,8 @@ import 'package:flutter/material.dart';
 import "package:intl/intl.dart";
 import 'package:photos/ente_theme_data.dart';
 import "package:photos/generated/l10n.dart";
+import 'package:photos/models/api/billing/subscription.dart';
 import "package:photos/models/api/storage_bonus/bonus.dart";
-import 'package:photos/models/subscription.dart';
 import "package:photos/theme/ente_theme.dart";
 import "package:photos/ui/components/captioned_text_widget.dart";
 import "package:photos/ui/components/menu_item_widget/menu_item_widget.dart";
