+++ mobile/lib/utils/delete_file_util.dart
@@ -13,10 +13,10 @@ import 'package:photos/events/files_updated_event.dart';
 import "package:photos/events/force_reload_trash_page_event.dart";
 import 'package:photos/events/local_photos_updated_event.dart';
 import "package:photos/generated/l10n.dart";
+import 'package:photos/models/api/collection/trash_item_request.dart';
 import 'package:photos/models/file/file.dart';
 import "package:photos/models/files_split.dart";
 import 'package:photos/models/selected_files.dart';
-import 'package:photos/models/trash_item_request.dart';
 import "package:photos/service_locator.dart";
 import 'package:photos/services/remote_sync_service.dart';
 import 'package:photos/services/sync_service.dart';
