+++ mobile/lib/utils/file_uploader.dart
@@ -29,8 +29,8 @@ import 'package:photos/models/encryption_result.dart';
 import 'package:photos/models/file/file.dart';
 import 'package:photos/models/file/file_type.dart';
 import "package:photos/models/metadata/file_magic.dart";
-import 'package:photos/models/upload_url.dart';
 import "package:photos/models/user_details.dart";
+import 'package:photos/module/upload/model/upload_url.dart';
 import "package:photos/module/upload/service/multipart.dart";
 import "package:photos/service_locator.dart";
 import 'package:photos/services/collections_service.dart';
