--- mobile/lib/utils/xml_parser_util.dart
@@ -1,1 +1,0 @@
-
